+++ homeassistant/components/lastfm/__init__.py
@@ -2,11 +2,8 @@
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers.device_registry import DeviceEntryType
-from homeassistant.helpers.entity import DeviceInfo
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from .const import DEFAULT_NAME, DOMAIN
+from .const import DOMAIN
 from .coordinator import LastFmUpdateCoordinator
 
 PLATFORMS = [Platform.SENSOR]
@@ -27,20 +24,3 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     if unload_ok := await hass.config_entries.async_unload_platforms(entry, PLATFORMS):
         hass.data[DOMAIN].pop(entry.entry_id)
     return unload_ok
-
-
-class LastFmEntity(CoordinatorEntity[LastFmUpdateCoordinator]):
-    """Representation of a LastFM entity."""
-
-    _attr_attribution = "Data provided by Last.fm"
-
-    def __init__(self, lastfm_coordinator: LastFmUpdateCoordinator) -> None:
-        """Initialize a LastFM entity."""
-        super().__init__(lastfm_coordinator)
-        self._attr_device_info = DeviceInfo(
-            configuration_url="https://www.last.fm",
-            entry_type=DeviceEntryType.SERVICE,
-            identifiers={(DOMAIN, lastfm_coordinator.config_entry.entry_id)},
-            manufacturer=DEFAULT_NAME,
-            name=DEFAULT_NAME,
-        )
